Move and Split Sources
