+++ src/main/webapp/components/sources-info/index.js
@@ -1,0 +1,2 @@
+import SourcesInfoGQL, { SourcesInfo } from './sources-info'
+export { SourcesInfoGQL as default, SourcesInfo }
