+++ src/main/webapp/components/sources-select/index.js
@@ -1,0 +1,4 @@
+import SourcesSelectGQL, {
+  SourcesSelectComponent as SourcesSelect,
+} from './sources-select'
+export { SourcesSelectGQL as default, SourcesSelect }
